+++ CHANGELOG.md
@@ -13,7 +13,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - added add_path to wex::factory::link
 - added separate styling for not yet committed lines
 - added vim gJ command
-- used boost::parser
 
 ### Changed
 
@@ -22,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - no longer use auto keyword in public interfaces
 - more stricter usage of setting ex marker
 - improved xml fold all
+- used boost::parser for processing git diff
 
 ### Fixed